Export the current external id version in tsv report
When needing to compare the current external id version associated with
an analysis record to the latest/newest external id version being
returned from a provider, we need to know what the current/previous
version is.

Previously in the tsv report, the latest/newest external id version was
returned. This change only modifies the tsv report to return the current
version. We want to return the latest version to shesmu so that it can
update vidarr's external id version if shesmu calculates a matching
signature (pinery-hash autoupdating feature).